requirements.txt
Biopython
pySBOL